#13 Merge module (combination → comparable)
