+++ ndtools/__init__.py
@@ -10,7 +10,6 @@ __all__ = [
     "TotalEquality",
     "TotalOrdering",
     "collections",
-    "combination",
     "comparable",
     "comparison",
     "operators",
@@ -20,11 +19,9 @@ __version__ = "0.3.0"
 
 # dependencies
 from . import collections
-from . import combination
 from . import comparable
 from . import comparison
 from . import operators
 from .collections import *
-from .combination import *
 from .comparable import *
 from .comparison import *
